Implement :pull_request_review_payload factory
Implement a factory with the common fields of the Github webhook pull_request_review payload.
As there is an independent factory for Models, when using Model and Payload factories make sure that if the ids or other values
must be in sync between Models and payloads they are built in sync.